+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added changelog
 - Added CHD8 mouse template (`/cpac_templates/chd8_functional_template_noise_mask_ag.nii.gz`) 
 - Added commandline flags `--T1w_label` and `--bold_label`
+- Added a step to store `desc-cleaned_bold` files with datatype `float`
 
 ### Changed